+++ packages/twenty-server/src/engine/core-modules/upgrade/decorators/registered-instance-command.decorator.ts
