+++ packages/twenty-server/src/engine/core-modules/upgrade/decorators/registered-workspace-command.decorator.ts
